+++ app/src/main/java/com/example/king_bob_nae/base/di/RetrofitModule.kt
@@ -1,0 +1,26 @@
+package com.example.king_bob_nae.base.di
+
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+class RetrofitModule {
+    @Provides
+    @Singleton
+    fun provideRetrofit(): Retrofit =
+        Retrofit.Builder()
+            .baseUrl(BASE_URL)
+            .addConverterFactory(GsonConverterFactory.create())
+            .build()
+
+
+    companion object {
+        private const val BASE_URL = "http://kki-log-api.turastory.com/"
+    }
+}
